+++ src/assets/scripts/client/aircraft/AircraftModel.js
@@ -4,8 +4,10 @@ import _forEach from 'lodash/forEach';
 import _get from 'lodash/get';
 import _isEqual from 'lodash/isEqual';
 import _isNil from 'lodash/isNil';
-import _round from 'lodash/round';
 import _uniqueId from 'lodash/uniqueId';
+import AirportController from '../airport/AirportController';
+import GameController, { GAME_EVENTS } from '../game/GameController';
+import UiController from '../UiController';
 import Fms from './FlightManagementSystem/Fms';
 import ModeController from './ModeControl/ModeController';
 import Pilot from './Pilot/Pilot';
@@ -52,7 +54,6 @@ import {
     AIRPORT_CONSTANTS,
     AIRPORT_CONTROL_POSITION_NAME
 } from '../constants/airportConstants';
-import { GAME_EVENTS } from '../game/GameController';
 import { MCP_MODE, MCP_MODE_NAME } from './ModeControl/modeControlConstants';
 import { TIME } from '../constants/globalConstants';
 
@@ -427,7 +428,7 @@ export default class AircraftModel {
         };
 
         this.takeoffTime = options.category === FLIGHT_CATEGORY.ARRIVAL
-            ? window.gameController.game_time()
+            ? GameController.game_time()
             : null;
 
         this.buildCurrentTerrainRanges();
@@ -513,7 +514,7 @@ export default class AircraftModel {
      * @method buildRestrictedAreaLinks
      */
     buildRestrictedAreaLinks() {
-        const restrictedAreas = prop.airport.current.restricted_areas;
+        const restrictedAreas = AirportController.current.restricted_areas;
 
         _forEach(restrictedAreas, (area) => {
             this.restricted.list.push({
@@ -543,7 +544,7 @@ export default class AircraftModel {
     }
 
     initFms(data) {
-        const airport = window.airportController.airport_get();
+        const airport = AirportController.airport_get();
         const initialRunway = airport.getActiveRunwayForCategory(this.category);
         this.fms = new Fms(data, initialRunway, this.model, this._navigationLibrary);
 
@@ -643,13 +644,13 @@ export default class AircraftModel {
                 AIRPORT_CONTROL_POSITION_NAME.DEPARTURE,
                 true
             );
-            window.gameController.events_recordNew(GAME_EVENTS.NOT_CLEARED_ON_ROUTE);
+            GameController.events_recordNew(GAME_EVENTS.NOT_CLEARED_ON_ROUTE);
 
             return;
         }
 
         this.radioCall('switching to center, good day', AIRPORT_CONTROL_POSITION_NAME.DEPARTURE);
-        window.gameController.events_recordNew(GAME_EVENTS.DEPARTURE);
+        GameController.events_recordNew(GAME_EVENTS.DEPARTURE);
     }
 
     /**
@@ -661,7 +662,7 @@ export default class AircraftModel {
     arrivalExit() {
         this.setIsRemovable();
         this.radioCall('leaving radar coverage as arrival', AIRPORT_CONTROL_POSITION_NAME.APPROACH, true);
-        window.gameController.events_recordNew(GAME_EVENTS.AIRSPACE_BUST);
+        GameController.events_recordNew(GAME_EVENTS.AIRSPACE_BUST);
     }
 
     /**
@@ -868,7 +869,7 @@ export default class AircraftModel {
      */
     isOnGround() {
         const errorAllowanceInFeet = 5;
-        const airport = window.airportController.airport_get();
+        const airport = AirportController.airport_get();
         const nearRunwayAltitude = abs(this.altitude - this.fms.currentRunway.elevation) < errorAllowanceInFeet;
         const nearAirportAltitude = abs(this.altitude - airport.elevation) < errorAllowanceInFeet;
 
@@ -961,7 +962,7 @@ export default class AircraftModel {
             head: 0
         };
 
-        const { wind } = window.airportController.airport_get();
+        const { wind } = AirportController.airport_get();
         // const wind = airport.wind;
         const angle = this.fms.currentRunway.calculateCrosswindAngleForRunway(wind.angle);
 
@@ -1004,19 +1005,19 @@ export default class AircraftModel {
         const callsign_S = this.getRadioCallsign();
 
         if (sectorType) {
-            call += window.airportController.airport_get().radio[sectorType];
+            call += AirportController.airport_get().radio[sectorType];
         }
 
         // call += ", " + this.callsign + " " + msg;
 
         // TODO: quick abstraction, this doesn't belong here.
-        const logMessage = (callsign) => `${window.airportController.airport_get().radio[sectorType]}, ${callsign} ${msg}`;
+        const logMessage = (callsign) => `${AirportController.airport_get().radio[sectorType]}, ${callsign} ${msg}`;
 
         if (alert) {
             const isWarning = true;
-            window.uiController.ui_log(logMessage(callsign_L), isWarning);
+            UiController.ui_log(logMessage(callsign_L), isWarning);
         } else {
-            window.uiController.ui_log(logMessage(callsign_L));
+            UiController.ui_log(logMessage(callsign_L));
         }
 
         speech_say([{
@@ -1050,18 +1051,18 @@ export default class AircraftModel {
                 alt_say = `at ${radio_altitude(alt)}`;
             }
 
-            window.uiController.ui_log(`${window.airportController.airport_get().radio.app}, ${this.callsign} with you ${alt_log}`);
+            UiController.ui_log(`${AirportController.airport_get().radio.app}, ${this.callsign} with you ${alt_log}`);
             speech_say([
-                { type: 'text', content: `${window.airportController.airport_get().radio.app}, ` },
+                { type: 'text', content: `${AirportController.airport_get().radio.app}, ` },
                 { type: 'callsign', content: this },
                 { type: 'text', content: `with you ${alt_say}` }
             ]);
         }
 
         if (this.category === FLIGHT_CATEGORY.DEPARTURE) {
-            window.uiController.ui_log(`${window.airportController.airport_get().radio.twr}, ${this.callsign}, ready to taxi`);
+            UiController.ui_log(`${AirportController.airport_get().radio.twr}, ${this.callsign}, ready to taxi`);
             speech_say([
-                { type: 'text', content: window.airportController.airport_get().radio.twr },
+                { type: 'text', content: AirportController.airport_get().radio.twr },
                 { type: 'callsign', content: this },
                 { type: 'text', content: ', ready to taxi' }
             ]);
@@ -1081,19 +1082,19 @@ export default class AircraftModel {
 
         // TODO: these two if blocks could be done in a single switch statement
         if (components.cross >= 20) {
-            window.gameController.events_recordNew(GAME_EVENTS.EXTREME_CROSSWIND_OPERATION);
-            window.uiController.ui_log(`${this.callsign} ${action} with major crosswind'`, isWarning);
+            GameController.events_recordNew(GAME_EVENTS.EXTREME_CROSSWIND_OPERATION);
+            UiController.ui_log(`${this.callsign} ${action} with major crosswind'`, isWarning);
         } else if (components.cross >= 10) {
-            window.gameController.events_recordNew(GAME_EVENTS.HIGH_CROSSWIND_OPERATION);
-            window.uiController.ui_log(`${this.callsign} ${action} with crosswind'`, isWarning);
+            GameController.events_recordNew(GAME_EVENTS.HIGH_CROSSWIND_OPERATION);
+            UiController.ui_log(`${this.callsign} ${action} with crosswind'`, isWarning);
         }
 
         if (components.head <= -10) {
-            window.gameController.events_recordNew(GAME_EVENTS.EXTREME_TAILWIND_OPERATION);
-            window.uiController.ui_log(`${this.callsign} ${action} with major tailwind'`, isWarning);
+            GameController.events_recordNew(GAME_EVENTS.EXTREME_TAILWIND_OPERATION);
+            UiController.ui_log(`${this.callsign} ${action} with major tailwind'`, isWarning);
         } else if (components.head <= -1) {
-            window.gameController.events_recordNew(GAME_EVENTS.HIGH_TAILWIND_OPERATION);
-            window.uiController.ui_log(`${this.callsign} ${action} with tailwind'`, isWarning);
+            GameController.events_recordNew(GAME_EVENTS.HIGH_TAILWIND_OPERATION);
+            UiController.ui_log(`${this.callsign} ${action} with tailwind'`, isWarning);
         }
 
         return score;
@@ -1250,7 +1251,7 @@ export default class AircraftModel {
 
         switch (this.flightPhase) {
             case FLIGHT_PHASE.TAXI: {
-                const elapsed = window.gameController.game_time() - this.taxi_start;
+                const elapsed = GameController.game_time() - this.taxi_start;
 
                 if (elapsed > this.taxi_time) {
                     this.setFlightPhase(FLIGHT_PHASE.WAITING);
@@ -1578,8 +1579,8 @@ export default class AircraftModel {
 
             const isWarning = true;
             // TODO: Should be moved to where the output is handled
-            window.gameController.events_recordNew(GAME_EVENTS.GO_AROUND);
-            window.uiController.ui_log(`${this.getRadioCallsign()} aborting landing, lost ILS`, isWarning);
+            GameController.events_recordNew(GAME_EVENTS.GO_AROUND);
+            UiController.ui_log(`${this.getRadioCallsign()} aborting landing, lost ILS`, isWarning);
 
             speech_say([
                 { type: 'callsign', content: this },
@@ -1596,8 +1597,8 @@ export default class AircraftModel {
     warnInterceptAngle() {
         const isWarning = true;
 
-        window.uiController.ui_log(`${this.callsign} approach course intercept angle was greater than 30 degrees`, isWarning);
-        window.gameController.events_recordNew(GAME_EVENTS.ILLEGAL_APPROACH_CLEARANCE);
+        UiController.ui_log(`${this.callsign} approach course intercept angle was greater than 30 degrees`, isWarning);
+        GameController.events_recordNew(GAME_EVENTS.ILLEGAL_APPROACH_CLEARANCE);
     }
 
     /**
@@ -1655,7 +1656,7 @@ export default class AircraftModel {
         const offset = getOffset(this, hold.fixPos, hold.inboundHeading);
         const holdLegDurationInSeconds = hold.legLength * TIME.ONE_MINUTE_IN_SECONDS;
         const bearingToHoldFix = vradial(vsub(hold.fixPos, this.relativePosition));
-        const gameTime = window.gameController.game.time;
+        const gameTime = GameController.game.time;
         const isPastFix = offset[1] < 1 && offset[2] < 2;
         const isTimerSet = hold.timer !== invalidTimerValue;
         const isTimerExpired = isTimerSet && gameTime > this.fms.currentWaypoint.timer;
@@ -1781,7 +1782,7 @@ export default class AircraftModel {
 
         if (this.hit) {
             // 90fps fall rate?...
-            this.altitude -= 90 * window.gameController.game_delta();
+            this.altitude -= 90 * GameController.game_delta();
             this.speed *= 0.99;
 
             return;
@@ -1803,11 +1804,11 @@ export default class AircraftModel {
             this.relativePositionHistory.push([
                 this.positionModel.relativePosition[0],
                 this.positionModel.relativePosition[1],
-                window.gameController.game_time() / window.gameController.game_speedup()
+                GameController.game_time() / GameController.game_speedup()
             ]);
             // TODO: this can be abstracted
-        } else if (abs((window.gameController.game_time() / window.gameController.game_speedup()) - this.relativePositionHistory[this.relativePositionHistory.length - 1][2]) > 4 / window.gameController.game_speedup()) {
-            this.relativePositionHistory.push([this.positionModel.relativePosition[0], this.positionModel.relativePosition[1], window.gameController.game_time() / window.gameController.game_speedup()]);
+        } else if (abs((GameController.game_time() / GameController.game_speedup()) - this.relativePositionHistory[this.relativePositionHistory.length - 1][2]) > 4 / GameController.game_speedup()) {
+            this.relativePositionHistory.push([this.positionModel.relativePosition[0], this.positionModel.relativePosition[1], GameController.game_time() / GameController.game_speedup()]);
         }
 
         this.updateGroundSpeedPhysics();
@@ -1816,7 +1817,7 @@ export default class AircraftModel {
         this.radial = radians_normalize(vradial(this.positionModel.relativePosition));
 
         // TODO: I am not sure what this has to do with aircraft Physics
-        const isInsideAirspace = this.isInsideAirspace(window.airportController.airport_get());
+        const isInsideAirspace = this.isInsideAirspace(AirportController.airport_get());
 
         if (isInsideAirspace !== this.inside_ctr) {
             this.crossBoundary(isInsideAirspace);
@@ -1836,7 +1837,7 @@ export default class AircraftModel {
             return;
         }
 
-        const secondsElapsed = window.gameController.game_delta();
+        const secondsElapsed = GameController.game_delta();
         const angle_diff = angle_offset(this.target.heading, this.heading);
         const angle_change = PERFORMANCE.TURN_RATE * secondsElapsed;
 
@@ -1894,7 +1895,7 @@ export default class AircraftModel {
         }
 
         const feetPerSecond = descentRate * TIME.ONE_SECOND_IN_MINUTES;
-        const feetDescended = feetPerSecond * window.gameController.game_delta();
+        const feetDescended = feetPerSecond * GameController.game_delta();
 
         if (abs(altitude_diff) < feetDescended) {
             this.altitude = this.target.altitude;
@@ -1920,7 +1921,7 @@ export default class AircraftModel {
         }
 
         const feetPerSecond = climbRate * TIME.ONE_SECOND_IN_MINUTES;
-        const feetClimbed = feetPerSecond * window.gameController.game_delta();
+        const feetClimbed = feetPerSecond * GameController.game_delta();
 
 
         if (abs(altitude_diff) < abs(feetClimbed)) {
@@ -1947,13 +1948,13 @@ export default class AircraftModel {
         }
 
         if (this.speed > this.target.speed) {
-            speedChange = -this.model.rate.decelerate * window.gameController.game_delta() / 2;
+            speedChange = -this.model.rate.decelerate * GameController.game_delta() / 2;
 
             if (this.isOnGround()) {
                 speedChange *= PERFORMANCE.DECELERATION_FACTOR_DUE_TO_GROUND_BRAKING;
             }
         } else if (this.speed < this.target.speed) {
-            speedChange  = this.model.rate.accelerate * window.gameController.game_delta() / 2;
+            speedChange  = this.model.rate.accelerate * GameController.game_delta() / 2;
             speedChange *= extrapolate_range_clamp(0, this.speed, this.model.speed.min, 2, 1);
         }
 
@@ -1982,7 +1983,7 @@ export default class AircraftModel {
 
         // Calculate wind vector
         const windIncreaseFactorPerFoot = 0.00002;  // 2.00% per thousand feet
-        const wind = window.airportController.airport_get().wind;
+        const wind = AirportController.airport_get().wind;
         const windTravelDirection = wind.angle + Math.PI;
         const windTravelSpeedAtSurface = wind.speed;
         const windTravelSpeed = windTravelSpeedAtSurface * (1 + (this.altitude * windIncreaseFactorPerFoot));
@@ -1994,7 +1995,7 @@ export default class AircraftModel {
         const groundSpeed = vlen(flightPathVector);
 
         // Calculate new position
-        const hoursElapsed = window.gameController.game_delta() * TIME.ONE_SECOND_IN_HOURS;
+        const hoursElapsed = GameController.game_delta() * TIME.ONE_SECOND_IN_HOURS;
         const distanceTraveled_nm = groundSpeed * hoursElapsed;
 
         this.positionModel.setCoordinatesByBearingAndDistance(groundTrack, distanceTraveled_nm);
@@ -2019,7 +2020,7 @@ export default class AircraftModel {
         //     // TODO: this should be abstracted to a helper function
         //     vector = vadd(vscale(
         //         vturn(wind.angle + Math.PI),
-        //         wind.speed * 0.000514444 * window.gameController.game_delta()),
+        //         wind.speed * 0.000514444 * GameController.game_delta()),
         //         vscale(vturn(angle + crab_angle), trueAirSpeed)
         //     );
         // }
@@ -2035,7 +2036,7 @@ export default class AircraftModel {
      * @deprecated
      */
     updateSimpleGroundSpeedPhysics() {
-        // const hoursElapsed = window.gameController.game_delta() * TIME.ONE_SECOND_IN_HOURS;
+        // const hoursElapsed = GameController.game_delta() * TIME.ONE_SECOND_IN_HOURS;
         // const distanceTraveled_nm = this.speed * hoursElapsed;
         //
         // this.positionModel.setCoordinatesByBearingAndDistance(this.heading, distanceTraveled_nm);
@@ -2099,7 +2100,7 @@ export default class AircraftModel {
 
                     // ac has just entered the area: .inside is still false, but st is true
                     if (new_inside && !area.inside) {
-                        window.gameController.events_recordNew(GAME_EVENTS.AIRSPACE_BUST);
+                        GameController.events_recordNew(GAME_EVENTS.AIRSPACE_BUST);
                         area.range = this.speed * 1.85 / 3.6 * 50 / 1000; // check in 50 seconds
                         // speed is kts, range is km.
                         // if a plane got into restricted area, don't check it too often
@@ -2122,7 +2123,7 @@ export default class AircraftModel {
         }
 
         if (this.terrain_ranges && !this.isOnGround()) {
-            const terrain = prop.airport.current.terrain;
+            const terrain = AirportController.current.terrain;
             const prev_level = this.terrain_ranges[this.terrain_level];
             const ele = Math.ceil(this.altitude, 1000);
             let curr_ranges = this.terrain_ranges[ele];
@@ -2151,13 +2152,13 @@ export default class AircraftModel {
 
                             console.log('hit terrain');
                             const isWarning = true;
-                            window.uiController.ui_log(`${this.callsign} collided with terrain in controlled flight`, isWarning);
+                            UiController.ui_log(`${this.callsign} collided with terrain in controlled flight`, isWarning);
                             speech_say([
                                 { type: 'callsign', content: this },
                                 { type: 'text', content: ', we\'re going down!' }
                             ]);
 
-                            window.gameController.events_recordNew(GAME_EVENTS.COLLISION);
+                            GameController.events_recordNew(GAME_EVENTS.COLLISION);
                         }
                     } else {
                         curr_ranges[id] = Math.max(0.2, status.distance);