+++ src/assets/scripts/client/trafficGenerator/SpawnPatternModel.js
@@ -5,6 +5,7 @@ import _isArray from 'lodash/isArray';
 import _isEmpty from 'lodash/isEmpty';
 import _random from 'lodash/random';
 import _round from 'lodash/round';
+import AirportController from '../airport/AirportController';
 import BaseModel from '../base/BaseModel';
 import StaticPositionModel from '../base/StaticPositionModel';
 import RouteModel from '../navigationLibrary/Route/RouteModel';
@@ -88,10 +89,9 @@ export default class SpawnPatternModel extends BaseModel {
      * @for SpawnPatternModel
      * @param spawnPatternJson {object}
      * @param navigationLibrary {NavigationLibrary}
-     * @param airportController {AirportController}
      */
     // istanbul ignore next
-    constructor(spawnPatternJson, navigationLibrary, airportController) {
+    constructor(spawnPatternJson, navigationLibrary) {
         super('spawnPatternModel');
 
         /**
@@ -107,19 +107,9 @@ export default class SpawnPatternModel extends BaseModel {
          * @property scheduleId
          * @type {number}
          * @default -1
-         * @property _airportController
-         * @type {AirportController}
          * @private
          */
         this.scheduleId = -1;
-
-        /**
-         *
-         * @property _airportController
-         * @type {AirportController}
-         * @private
-         */
-        this._airportController = null;
 
         /**
          * One of `FLIGHT_CATEGORY`
@@ -405,7 +395,7 @@ export default class SpawnPatternModel extends BaseModel {
          */
         this._uptime = -1;
 
-        this.init(spawnPatternJson, navigationLibrary, airportController);
+        this.init(spawnPatternJson, navigationLibrary);
     }
 
     /**
@@ -466,9 +456,8 @@ export default class SpawnPatternModel extends BaseModel {
      * @method init
      * @param spawnPatternJson {object}
      * @param navigationLibrary {NavigationLibrary}
-     * @param airportController {AirportController}
      */
-    init(spawnPatternJson, navigationLibrary, airportController) {
+    init(spawnPatternJson, navigationLibrary) {
         // We return early here if the object is empty because we pre-hydrate objects in the `ModelSourcePool`
         if (_isEmpty(spawnPatternJson)) {
             return;
@@ -478,16 +467,11 @@ export default class SpawnPatternModel extends BaseModel {
             throw new TypeError('Invalid NavigationLibrary passed to SpawnPatternModel');
         }
 
-        if (isEmptyObject(airportController)) {
-            throw new TypeError('Invalid AirportController passed to SpawnPatternModel');
-        }
-
         // TODO: this is a temporary development check. this should be removed before merging in to develop
         if (!spawnPatternModelJsonValidator(spawnPatternJson)) {
             console.error('### Invalid spawnPatternJson received', spawnPatternJson);
         }
 
-        this._airportController = airportController;
         this.origin = spawnPatternJson.origin;
         this.destination = spawnPatternJson.destination;
         this.category = spawnPatternJson.category;
@@ -502,7 +486,7 @@ export default class SpawnPatternModel extends BaseModel {
         this._maximumDelay = this._calculateMaximumDelayFromSpawnRate();
         this.airlines = this._assembleAirlineNamesAndFrequencyForSpawn(spawnPatternJson.airlines);
         this._weightedAirlineList = this._buildWeightedAirlineList();
-        this.preSpawnAircraftList = this._buildPreSpawnAircraft(spawnPatternJson, navigationLibrary, airportController);
+        this.preSpawnAircraftList = this._buildPreSpawnAircraft(spawnPatternJson, navigationLibrary);
 
         this._calculateSurgePatternInitialDelayValues(spawnPatternJson);
         this._setCyclePeriodAndOffset(spawnPatternJson);
@@ -968,9 +952,8 @@ export default class SpawnPatternModel extends BaseModel {
      * @method _buildPreSpawnAircraft
      * @param spawnPatternJson {object}
      * @param navigationLibrary {NavigationLibrary}
-     * @param airportController {AirportController}
      */
-    _buildPreSpawnAircraft(spawnPatternJson, navigationLibrary, airportController) {
+    _buildPreSpawnAircraft(spawnPatternJson, navigationLibrary) {
         if (this._isDeparture()) {
             // TODO: this may be dead, please remove if it is
             const preSpawnDepartureAircraft = [{
@@ -983,7 +966,7 @@ export default class SpawnPatternModel extends BaseModel {
         const preSpawnArrivalAircraftList = buildPreSpawnAircraft(
             spawnPatternJson,
             navigationLibrary,
-            airportController.airport.current
+            AirportController.current
         );
 
         return preSpawnArrivalAircraftList;
@@ -1015,7 +998,6 @@ export default class SpawnPatternModel extends BaseModel {
         this.heading = heading;
     }
 
-    // TODO: add support for `FIXXA..FIXXB.PRCDR1.FIXXC` route strings
     /**
      * Given a `routeString`, find the `FixModel`s or `WaypointModel`s associated with that route.
      *
@@ -1043,7 +1025,7 @@ export default class SpawnPatternModel extends BaseModel {
         const waypointModelList = navigationLibrary.starCollection.findRouteWaypointsForRouteByEntryAndExit(
             routeModel.procedure,
             routeModel.entry,
-            this._airportController.getInitialArrivalRunwayName(),
+            AirportController.getInitialArrivalRunwayName(),
             isPreSpawn
         );
 
@@ -1057,7 +1039,7 @@ export default class SpawnPatternModel extends BaseModel {
      * @return {StaticPositionModel}
      */
     _generateSelfReferencedAirportPositionModel() {
-        const airportPosition = this._airportController.airport_get().positionModel;
+        const airportPosition = AirportController.airport_get().positionModel;
         const selfReferencingPosition = new StaticPositionModel(
             airportPosition.gps,
             airportPosition,